Pictures of the Toy
These are pictures of the inner workings of the toy. They are messy and hot glued to hell but I didn't think a cheap toy such as this one warranted 3d printed parts or anything else to elaborate. I used fishing line (20lb Test) and a pencil to rig the keys to the servos (20201231_090508.jpg). I had to cut some of the key arms to allow room for the servos as you can see in the picture (20201231_090531.jpg). The 9 volt battery is also cheaply rigged to a switch and the Prototyping board to power the Arduino Nano I could not find and LIPO batteries in my area.